backend/requirements.txt
flask==3.0.3
flask-cors==4.0.1
sqlalchemy==2.0.31
psycopg2-binary==2.9.9
pillow==10.4.0
pytesseract==0.3.13
openai==1.40.0
alembic==1.13.2
python-dotenv==1.0.1
werkzeug==3.0.3
gunicorn==23.0.0
